Pull out an independent `symsrv` crate
