+++ crates/bin/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "pdblister"
+description = "Faster version of `symchk /om` for generating PDB manifests of offline machines"
+license-file = "LICENSE"
+homepage = "https://github.com/microsoft/pdblister"
+repository = "https://github.com/microsoft/pdblister"
+version = "0.0.2"
+authors = ["Gogs <gogs@fake.local>", "Justin Moore <jusmoore@microsoft.com>"]
+edition = "2018"
+
+[dependencies]
+anyhow = "1.0"
+clap = "4.0.22"
+futures = "0.3"
+indicatif = { version = "0.17.1", features = ["tokio"] }
+rand = "0.8"
+serde_json = "1.0.87"
+zerocopy = "0.6"
+
+symsrv = { path = "../symsrv" }
+
+[dependencies.tokio]
+version = "1.21.2"
+features = ["full"]
