+++ crates/bin/src/pe.rs
