+++ crates/symsrv/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "symsrv"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow = "1.0.66"
+base64 = "0.13"
+futures = "0.3.25"
+indicatif = "0.17.2"
+mime = "0.3"
+reqwest = "0.11.13"
+thiserror = "1.0.37"
+url = "2.2"
+
+[dependencies.tokio]
+version = "1.21.2"
+features = ["fs", "rt"]
